Changed Entity to be a 64 bit type instead of 32. This allows us to store which world the entity belongs to (8 bits). We have 24 bits in reserve at the moment. Changed name of AllocateEntity to AllocatedEntityId for clarity of what the function actually does. World hash is now a custom ID type, and worlds have revieved a world ID as well (index into world array). Fixed issue with entity generation id sometimes being wrong after converting from integers.

+++ code/application/game/componentserialization.cc
@@ -113,6 +113,6 @@ IO::JsonReader::Get<Game::Entity>(Game::Entity& entity, const char* key)
 template<> void
 IO::JsonWriter::Add<Game::Entity>(Game::Entity const& entity, Util::String const& key)
 {
-    this->Add<uint>((uint)entity, key);
+    this->Add<uint64_t>((uint64_t)entity, key);
 }
